+++ vortex-file/src/v2/file.rs
@@ -79,42 +79,3 @@ impl<R: VortexReadAt + Unpin> VortexFile<R> {
         Ok(ArrayStreamAdapter::new(result_dtype, stream))
     }
 }
-
-#[cfg(test)]
-mod test {
-    use bytes::Bytes;
-    use vortex_array::array::ChunkedArray;
-    use vortex_array::stream::ArrayStreamExt;
-    use vortex_array::{ContextRef, IntoArrayData, IntoArrayVariant};
-    use vortex_buffer::buffer;
-    use vortex_error::VortexResult;
-    use vortex_scan::Scan;
-
-    use crate::v2::*;
-
-    #[async_std::test]
-    async fn basic_file_roundtrip() -> VortexResult<()> {
-        let array = ChunkedArray::from_iter([
-            buffer![0, 1, 2].into_array(),
-            buffer![3, 4, 5].into_array(),
-            buffer![6, 7, 8].into_array(),
-        ])
-        .into_array();
-
-        let buffer: Bytes = WriteOptions::default()
-            .write_async(vec![], array.into_array_stream())
-            .await?
-            .into();
-
-        let vxf = OpenOptions::new(ContextRef::default()).open(buffer).await?;
-        let result = vxf
-            .scan(Scan::all())?
-            .into_array_data()
-            .await?
-            .into_primitive()?;
-
-        assert_eq!(result.as_slice::<i32>(), &[0, 1, 2, 3, 4, 5, 6, 7, 8]);
-
-        Ok(())
-    }
-}
