+++ vortex-file/src/v2/tests.rs
@@ -3,37 +3,33 @@ use vortex_array::array::ChunkedArray;
 use vortex_array::stream::ArrayStreamExt;
 use vortex_array::{ContextRef, IntoArrayData, IntoArrayVariant};
 use vortex_buffer::buffer;
+use vortex_error::VortexResult;
 use vortex_scan::Scan;
 
-use crate::v2::{OpenOptions, WriteOptions};
+use crate::v2::*;
 
-#[tokio::test]
-async fn write_read() {
-    let arr = ChunkedArray::from_iter(vec![
+#[async_std::test]
+async fn basic_file_roundtrip() -> VortexResult<()> {
+    let array = ChunkedArray::from_iter([
         buffer![0, 1, 2].into_array(),
         buffer![3, 4, 5].into_array(),
+        buffer![6, 7, 8].into_array(),
     ])
     .into_array();
 
-    let written = WriteOptions::default()
-        .write_async(vec![], arr.into_array_stream())
-        .await
-        .unwrap();
+    let buffer: Bytes = WriteOptions::default()
+        .write_async(vec![], array.into_array_stream())
+        .await?
+        .into();
 
-    // TODO(ngates): no need to wrap Vec<u8> in Bytes if VortexReadAt doesn't require clone.
-    let vxf = OpenOptions::new(ContextRef::default())
-        .open(Bytes::from(written))
-        .await
-        .unwrap();
-
+    let vxf = OpenOptions::new(ContextRef::default()).open(buffer).await?;
     let result = vxf
-        .scan(Scan::all())
-        .unwrap()
+        .scan(Scan::all())?
         .into_array_data()
-        .await
-        .unwrap()
-        .into_primitive()
-        .unwrap();
+        .await?
+        .into_primitive()?;
 
-    assert_eq!(result.as_slice::<i32>(), &[0, 1, 2, 3, 4, 5]);
+    assert_eq!(result.as_slice::<i32>(), &[0, 1, 2, 3, 4, 5, 6, 7, 8]);
+
+    Ok(())
 }
